requirements.txt
python=3.10.9
pytorch=1.12.1
torchvision=0.13.1
pillow=9.4.0
matplotlib=3.7.1
flask=2.2.2
python-dotenv=0.21.0